+++ development/backend/authentication/main/ioc/ioc/auth.ioc.ts
@@ -2,7 +2,7 @@ import cors from 'cors';
 import express from 'express';
 import { injectable, InjectionToken, singleton } from 'tsyringe';
 import { RegularUserIoC } from '../../../regular-user/main/ioc/ioc/regular-user.ioc';
-import { ModuleIoC } from '../../../shared/main/ioc/ioc/ioc-initializer.base';
+import { ModuleIoC } from '../../../shared/main/ioc/ioc/ioc';
 import { IAuthenticationConfig } from '../../config/auth.config.interface';
 import { AuthenticationBinder } from '../binders/auth.binder';
