+++ development/backend/authentication/regular-user/login/application/ports/config/login.config.interface.ts
@@ -1,1 +1,1 @@
-export interface IConfigLoginApplication {}
+export interface ILoginApplicationConfig {}
