+++ development/backend/authentication/regular-user/login/core/config/login.config.interface.ts
@@ -1,1 +1,1 @@
-export interface IConfigLoginCore {}
+export interface ILoginCoreConfig {}
