+++ development/backend/authentication/regular-user/login/main/config/login.config.interface.ts
@@ -1,9 +1,9 @@
-import { IConfigLoginApplication } from '../../application/ports/config/login.config.interface';
-import { IConfigLoginCore } from '../../core/config/login.config.interface';
+import { ILoginApplicationConfig } from '../../application/ports/config/login.config.interface';
+import { ILoginCoreConfig } from '../../core/config/login.config.interface';
 import { ILoginInfrastructureConfig } from '../../infrastructure/config/login.config.interface';
 
 export interface ILoginConfig {
-    readonly core: IConfigLoginCore;
-    readonly application: IConfigLoginApplication;
+    readonly core: ILoginCoreConfig;
+    readonly application: ILoginApplicationConfig;
     readonly infrastructure: ILoginInfrastructureConfig;
 }
