+++ development/backend/authentication/regular-user/login/main/ioc/ioc/login.ioc.ts
@@ -1,5 +1,5 @@
 import { injectable, InjectionToken, singleton } from 'tsyringe';
-import { FeatureIoC } from '../../../../../shared/main/ioc/ioc/ioc-initializer.base';
+import { FeatureIoC } from '../../../../../shared/main/ioc/ioc/ioc';
 import { LoginRouter } from '../../../infrastructure/api/routes/login.router';
 import { ILoginConfig } from '../../config/login.config.interface';
 import { LoginBinder } from '../binders/login.binder';
